Enum schemamama::Error

pub enum Error<E> {
    Adapter(E),
    Migration {
        version: Version,
        description: &'static str,
        direction: Direction,
        error: E,
    },
}

An all-encompassing error type that can be returned during interaction with the migrator adapter.

Variants

Adapter

A generic error that occurred while interacting with the adapter.

Migration

An error that arose from the adapter specifically during a migration's execution.

Fields

version

The version of the migration that failed.

description

The description of the migration that failed.

direction

The direction in which the failed migration was ran.

error

The underlying error from the adapter.